requirements.txt
gevent
Flask==2.1.2
DateTime==4.5
Werkzeug==2.1.2
gunicorn==19.9.0
Flask-Cors==3.0.10
flask-restx==0.5.1
SQLAlchemy==1.4.40
Flask-RESTful==0.3.9
python-dotenv==0.20.0
python-dateutil==2.8.2
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.9
shortuuid
validators
flask-login
jinja2 
weasyprint
Flask-Mail
APScheduler
pandas